# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Academic Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  detail?: string | null;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface DocumentUploadInput {
  file: Blob;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  pending: 'pending',
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface Document {
  id: string;
  userId: string;
  name: string;
  mimeType: string;
  size: number;
  status: DocumentStatus;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  wordCount?: number | null;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
}

export type SummarizeInputSummaryType = typeof SummarizeInputSummaryType[keyof typeof SummarizeInputSummaryType];


export const SummarizeInputSummaryType = {
  short: 'short',
  detailed: 'detailed',
  bullet: 'bullet',
  executive: 'executive',
} as const;

export interface SummarizeInput {
  summaryType: SummarizeInputSummaryType;
}

export type SummarySummaryType = typeof SummarySummaryType[keyof typeof SummarySummaryType];


export const SummarySummaryType = {
  short: 'short',
  detailed: 'detailed',
  bullet: 'bullet',
  executive: 'executive',
} as const;

export interface Summary {
  id: string;
  documentId: string;
  summaryType: SummarySummaryType;
  content: string;
  createdAt: string;
}

export type ConceptConceptType = typeof ConceptConceptType[keyof typeof ConceptConceptType];


export const ConceptConceptType = {
  topic: 'topic',
  keyword: 'keyword',
  definition: 'definition',
  formula: 'formula',
  date: 'date',
  person: 'person',
} as const;

export interface Concept {
  id: string;
  documentId: string;
  term: string;
  definition: string;
  conceptType: ConceptConceptType;
}

export interface ChatSessionInput {
  documentId: string;
  /** @nullable */
  title?: string | null;
}

export interface ChatSession {
  id: string;
  documentId: string;
  userId: string;
  title: string;
  messageCount: number;
  /** @nullable */
  documentName?: string | null;
  createdAt: string;
}

export interface ChatSource {
  content: string;
  /** @nullable */
  pageNumber: number | null;
  relevanceScore: number;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  sessionId: string;
  role: ChatMessageRole;
  content: string;
  sources?: ChatSource[];
  createdAt: string;
}

export interface ChatSessionWithMessages {
  id: string;
  documentId: string;
  userId: string;
  title: string;
  messageCount: number;
  /** @nullable */
  documentName?: string | null;
  createdAt: string;
  messages: ChatMessage[];
}

export interface MessageInput {
  content: string;
}

export type QuizInputDifficulty = typeof QuizInputDifficulty[keyof typeof QuizInputDifficulty];


export const QuizInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type QuizInputQuestionTypesItem = typeof QuizInputQuestionTypesItem[keyof typeof QuizInputQuestionTypesItem];


export const QuizInputQuestionTypesItem = {
  multiple_choice: 'multiple_choice',
  true_false: 'true_false',
  short_answer: 'short_answer',
} as const;

export interface QuizInput {
  difficulty: QuizInputDifficulty;
  /**
     * @minimum 1
     * @maximum 20
     */
  questionCount: number;
  questionTypes?: QuizInputQuestionTypesItem[];
}

export interface QuizOption {
  label: string;
  text: string;
}

export type QuizQuestionQuestionType = typeof QuizQuestionQuestionType[keyof typeof QuizQuestionQuestionType];


export const QuizQuestionQuestionType = {
  multiple_choice: 'multiple_choice',
  true_false: 'true_false',
  short_answer: 'short_answer',
} as const;

export interface QuizQuestion {
  id: string;
  question: string;
  questionType: QuizQuestionQuestionType;
  options?: QuizOption[];
  answer: string;
  /** @nullable */
  explanation?: string | null;
}

export type QuizDifficulty = typeof QuizDifficulty[keyof typeof QuizDifficulty];


export const QuizDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Quiz {
  id: string;
  documentId: string;
  /** @nullable */
  documentName?: string | null;
  difficulty: QuizDifficulty;
  questions: QuizQuestion[];
  /** @nullable */
  score?: number | null;
  /** @nullable */
  maxScore?: number | null;
  createdAt: string;
}

export type QuizSubmissionAnswers = {[key: string]: string};

export interface QuizSubmission {
  answers: QuizSubmissionAnswers;
}

export interface QuizFeedbackItem {
  questionId: string;
  correct: boolean;
  /** @nullable */
  yourAnswer?: string | null;
  correctAnswer: string;
  /** @nullable */
  explanation?: string | null;
}

export interface QuizResult {
  quizId: string;
  score: number;
  maxScore: number;
  percentage: number;
  feedback: QuizFeedbackItem[];
}

export interface FlashcardsInput {
  /**
     * @minimum 1
     * @maximum 50
     */
  cardCount: number;
  /** @nullable */
  topic?: string | null;
}

export interface Flashcard {
  id: string;
  question: string;
  answer: string;
}

export interface FlashcardSet {
  id: string;
  documentId: string;
  /** @nullable */
  documentName?: string | null;
  /** @nullable */
  topic?: string | null;
  cards: Flashcard[];
  createdAt: string;
}

export interface DashboardStats {
  totalDocuments: number;
  totalChats: number;
  totalQuizzes: number;
  totalFlashcardSets: number;
  totalFlashcards: number;
  documentsReady: number;
  /** @nullable */
  averageQuizScore?: number | null;
  recentDocuments: Document[];
}

export type ActivityItemActivityType = typeof ActivityItemActivityType[keyof typeof ActivityItemActivityType];


export const ActivityItemActivityType = {
  upload: 'upload',
  chat: 'chat',
  quiz: 'quiz',
  flashcard: 'flashcard',
  summary: 'summary',
} as const;

export interface ActivityItem {
  id: string;
  activityType: ActivityItemActivityType;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  documentId?: string | null;
  createdAt: string;
}