# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Academic Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Documents ───────────────────────────────────────────────────────────────
...
  /chat/sessions:
    get:
      operationId: listChatSessions
      tags: [chat]
      summary: List all chat sessions for current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session for a document
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{sessionId}:
    get:
      operationId: getChatSession
      tags: [chat]
      summary: Get a chat session with its messages
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionWithMessages"
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /chat/sessions/{sessionId}/messages:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to a chat session (RAG)
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # ─── Quiz ─────────────────────────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: ["string", "null"]
      required:
        - error

    # ─── Auth ──────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ─── Documents ─────────────────────────────────────────────────────────────
...
    Document:
      type: object
      required: [id, userId, name, mimeType, size, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        status:
          type: string
          enum: [pending, processing, ready, error]
        pageCount:
          type: ["integer", "null"]
        wordCount:
          type: ["integer", "null"]
        summary:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── Summaries ─────────────────────────────────────────────────────────────
...
    ChatSessionInput:
      type: object
      required: [documentId]
      properties:
        documentId:
          type: string
        title:
          type: ["string", "null"]
    ChatSession:
      type: object
      required: [id, documentId, userId, title, messageCount, createdAt]
      properties:
        id:
          type: string
        documentId:
          type: string
        userId:
          type: string
        title:
          type: string
        messageCount:
          type: integer
        documentName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ChatSource:
      type: object
      required: [content, pageNumber, relevanceScore]
      properties:
        content:
          type: string
        pageNumber:
          type: ["integer", "null"]
        relevanceScore:
          type: number
    ChatMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/ChatSource"
        createdAt:
          type: string
          format: date-time
    ChatSessionWithMessages:
      type: object
      required: [id, documentId, userId, title, messageCount, createdAt, messages]
      properties:
        id:
          type: string
        documentId:
          type: string
        userId:
          type: string
        title:
          type: string
        messageCount:
          type: integer
        documentName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ─── Quiz ──────────────────────────────────────────────────────────────────
...
    DashboardStats:
      type: object
      required:
        - totalDocuments
        - totalChats
        - totalQuizzes
        - totalFlashcardSets
        - totalFlashcards
        - documentsReady
        - recentDocuments
      properties:
        totalDocuments:
          type: integer
        totalChats:
          type: integer
        totalQuizzes:
          type: integer
        totalFlashcardSets:
          type: integer
        totalFlashcards:
          type: integer
        documentsReady:
          type: integer
        averageQuizScore:
          type: ["number", "null"]
        recentDocuments:
          type: array
          items:
            $ref: "#/components/schemas/Document"
    ActivityItem:
      type: object
      required: [id, activityType, title, createdAt]
      properties:
        id:
          type: string
        activityType:
          type: string
          enum: [upload, chat, quiz, flashcard, summary]
        title:
          type: string
        description:
          type: ["string", "null"]
        documentId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time