# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Academic Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemActivityType } from './activityItemActivityType';

export interface ActivityItem {
  id: string;
  activityType: ActivityItemActivityType;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  documentId?: string | null;
  createdAt: Date;
}