# lib/api-zod/src/generated/types/chatSessionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Academic Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSessionInput {
  documentId: string;
  /** @nullable */
  title?: string | null;
}