# lib/api-zod/src/generated/types/chatSessionWithMessages.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Academic Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatSessionWithMessages {
  id: string;
  documentId: string;
  userId: string;
  title: string;
  messageCount: number;
  /** @nullable */
  documentName?: string | null;
  createdAt: Date;
  messages: ChatMessage[];
}